# docs/design/openapi_sin.yaml
openapi: 3.0.3
info:
  title: Signal Ingestion Normalization Stub
  version: "1.0.0"
servers:
  - url: http://localhost
paths:
  /v1/health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  timestamp:
                    type: string
  /v1/ready:
    get:
      summary: Readiness
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ready:
                    type: boolean
                  checks:
                    type: object
                  timestamp:
                    type: string
  /v1/producers/register:
    post:
      summary: Register producer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                producer:
                  $ref: '#/components/schemas/ProducerRegistration'
              required: [producer]
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  producer_id:
                    type: string
                  status:
                    type: string
  /v1/producers:
    get:
      summary: List producers
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  producers:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProducerRegistration'
  /v1/producers/{producer_id}:
    parameters:
      - name: producer_id
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get producer
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  producer:
                    $ref: '#/components/schemas/ProducerRegistration'
    put:
      summary: Update producer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                producer:
                  $ref: '#/components/schemas/ProducerRegistration'
              required: [producer]
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  producer_id:
                    type: string
                  status:
                    type: string
  /v1/signals/ingest:
    post:
      summary: Ingest signals
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                signals:
                  type: array
                  minItems: 1
                  maxItems: 1000
                  items:
                    $ref: '#/components/schemas/SignalEnvelope'
              required: [signals]
      responses:
        '200':
          description: Ingest result
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                  summary:
                    type: object
  /v1/signals/dlq:
    get:
      summary: Inspect DLQ
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: tenant_id
          schema:
            type: string
        - in: query
          name: producer_id
          schema:
            type: string
        - in: query
          name: signal_type
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: offset
          schema:
            type: integer
      responses:
        '200':
          description: DLQ entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items:
                      type: object
                  total:
                    type: integer

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ProducerRegistration:
      type: object
      properties:
        producer_id:
          type: string
        name:
          type: string
        plane:
          type: string
          enum: [edge, cloud]
        owner:
          type: string
        allowed_signal_kinds:
          type: array
          items:
            type: string
            enum: [event, metric]
        allowed_signal_types:
          type: array
          items:
            type: string
        contract_versions:
          type: object
          additionalProperties:
            type: string
      required: [producer_id, name, plane, allowed_signal_kinds, allowed_signal_types, contract_versions]
    SignalEnvelope:
      type: object
      properties:
        signal_id:
          type: string
        tenant_id:
          type: string
        environment:
          type: string
          enum: [dev, prod, stage]
        producer_id:
          type: string
        signal_kind:
          type: string
          enum: [event, metric]
        signal_type:
          type: string
        occurred_at:
          type: string
          format: date-time
        ingested_at:
          type: string
          format: date-time
        payload:
          type: object
        schema_version:
          type: string
      required: [signal_id, tenant_id, environment, producer_id, signal_kind, signal_type, occurred_at, ingested_at, payload, schema_version]